feat(toast, error): add method to append toast and modified ajax methods to return errors
Made use of ajax error handling to return errors in case the request dont go through, those errors will be appended as a toast error message. Also modified suggestion divs, added a class to them to be able to delete them in bulk with the .remove clause. Also indide each toast we listen for a botstrap method called on hide and when it occurs we delete the object from DOM